pattern/file: move File under pattern/file/serialized
